+++ client/src/data/dummyData.js
@@ -1,10 +1,8 @@
-// This object mimics the full response from the Catalog API.
-// It is a valid JavaScript object.
+// src/data/dummyData.js
 export const dummyCatalogData = {
   respId: "RES00001",
   userType: "3",
   master: {
-    // 1. Dummy "Offer" data
     offer: [
       {
         name: "Random Blood Sugar",
@@ -13,38 +11,18 @@ export const dummyCatalogData = {
         childs: [{ name: "RBS", code: "RBS", groupName: "WELLNESS", type: "TEST" }],
         rate: { b2B: "", b2C: "280", offerRate: "99", id: "PROJ1020761" },
         testCount: "1",
-        benMin: "1",
-        benMax: "10",
-        payType: "PREPAID",
-        fasting: "NF", // No Fasting
+        fasting: "NF",
         imageLocation: "https://b2capi.thyrocare.com/wellness/img/Banner/RBS_banner.jpg",
       },
-      {
-        name: "Basic Allergy Test",
-        code: "PROJ11111",
-        type: "OFFER",
-        childs: [{ name: "ALLERGY", code: "ALG", groupName: "WELLNESS", type: "TEST" }],
-        rate: { b2B: "500", b2C: "1000", offerRate: "799", id: "PROJ11111" },
-        testCount: "1",
-        benMin: "1",
-        benMax: "10",
-        payType: "PREPAID",
-        fasting: "NF",
-        imageLocation: "https://via.placeholder.com/400x200?text=Allergy+Offer",
-      },
     ],
-    // 2. Dummy "Profile" data (which you call Packages)
     profile: [
       {
         name: "ADVANCED RENAL PROFILE",
         code: "P522",
         type: "PROFILE",
-        childs: [{ name: "BUN/SR.CREATININE RATIO", code: "B/CR", type: "TEST" }],
         rate: { b2B: "310", b2C: "650", offerRate: "650", id: "9821" },
         testCount: "11",
-        benMin: "1",
-        benMax: "10",
-        fasting: "CF", // Compulsory Fasting
+        fasting: "CF",
         specimenType: "SERUM",
         imageLocation: "https://b2capi.thyrocare.com/wellness/img/Banner/P522_banner.jpg",
       },
@@ -60,18 +38,14 @@ export const dummyCatalogData = {
         imageLocation: "http://b2capi.thyrocare.com/wellness/img/Banner/P187_banner.jpg",
       },
     ],
-    // 3. Dummy "Test" data
     tests: [
       {
         name: "FASTING BLOOD SUGAR",
         code: "FBS",
         type: "TEST",
-        childs: [],
         rate: { b2B: "15", b2C: "149", offerRate: "149", id: "11313" },
         testCount: "1",
-        benMin: "1",
-        benMax: "10",
-        fasting: "CF", // Compulsory Fasting
+        fasting: "CF",
         specimenType: "FLUORIDE",
         imageLocation: "https://b2capi.thyrocare.com/wellness/img/Banner/FBS_banner.jpg",
       },
@@ -79,11 +53,8 @@ export const dummyCatalogData = {
         name: "LIPID PROFILE",
         code: "LIPID",
         type: "TEST",
-        childs: [],
         rate: { b2B: "200", b2C: "500", offerRate: "449", id: "12345" },
         testCount: "8",
-        benMin: "1",
-        benMax: "10",
         fasting: "CF",
         specimenType: "SERUM",
         imageLocation: "https://b2capi.thyrocare.com/wellness/img/Banner/LIPID_banner.jpg",
