+++ client/src/pages/LandingPage.jsx
@@ -1,47 +1,40 @@
 // src/pages/LandingPage.jsx
-
 import React from 'react';
-import { dummyCatalogData } from '../components/data/dummyData'; // Import the dummy data
+import { dummyCatalogData } from '../data/dummyData'; // Import the dummy data
 
 // Import sections
-import Hero from '../components/sections/Hero';
-import HomeCarousel from '../components/sections/HomeCarousel';
-import ProductGrid from '../components/sections/ProductGrid';
+import Hero from '../sections/Hero';
+import HomeCarousel from '../sections/HomeCarousel';
+import ProductGrid from '../sections/ProductGrid';
 
 const LandingPage = () => {
-  // [cite_start]// Extract the arrays from the dummy data [cite: 214-221]
   const offers = dummyCatalogData.master.offer;
-  const packages = dummyCatalogData.master.profile; // "Profile" is your "Package"
+  const packages = dummyCatalogData.master.profile;
   const tests = dummyCatalogData.master.tests;
 
   return (
     <div className="max-w-7xl mx-auto px-4 sm:px-6 py-10">
       
-      {/* 1. Hero Section (Photo + Text) */}
       <Hero />
       
-      {/* 2. Carousel Section */}
       <HomeCarousel />
       
-      {/* 3. Offer Section */}
       <ProductGrid 
         title="Top Offers" 
         products={offers} 
-        seeAllLink="/offers" 
+        seeAllLink="#" 
       />
       
-      {/* 4. Package (Profile) Section */}
       <ProductGrid 
         title="Popular Packages" 
         products={packages} 
-        seeAllLink="/packages" 
+        seeAllLink="#" 
       />
       
-      {/* 5. Test Section (as seen in your sketch) */}
       <ProductGrid 
         title="Featured Tests" 
         products={tests} 
-        seeAllLink="/tests" 
+        seeAllLink="#" 
       />
 
     </div>
